Move _unwrap_ and _expect_ to next to the `Undefinable` type definition
