+++ src/Undefinable/index.ts
@@ -1,9 +1,15 @@
-export { NotUndefined, Undefinable, isNotUndefined, isUndefined } from './Undefinable.js';
+export {
+    type NotUndefined,
+    type Undefinable,
+    isNotUndefined,
+    isUndefined,
+    expectNotUndefined,
+    unwrapUndefinable,
+} from './Undefinable.js';
 // XXX: `and()` operation is equivalent of `a && b` so we don't ship it by default set.
 //export { andForUndefinable as and} from './and.js';
 export { andThenForUndefinable as andThen } from './andThen.js';
 export { andThenAsyncForUndefinable as andThenAsync } from './andThenAsync.js';
-export { expectNotUndefined as expect } from './expect.js';
 export { inspectUndefinable as inspect } from './inspect.js';
 export { mapForUndefinable as map } from './map.js';
 export { mapAsyncForUndefinable as mapAsync } from './mapAsync.js';
@@ -15,7 +21,22 @@ export { mapOrElseAsyncForUndefinable as mapOrElseAsync } from './mapOrElseAsync
 // export { orForUndefinable as or} from './or.js';
 export { orElseForUndefinable as orElse } from './orElse.js';
 export { orElseAsyncForUndefinable as orElseAsync } from './orElseAsync.js';
-export { unwrapUndefinable as unwrap } from './unwrap.js';
 export { unwrapOrFromUndefinable as unwrapOr } from './unwrapOr.js';
 export { unwrapOrElseFromUndefinable as unwrapOrElse } from './unwrapOrElse.js';
 export { unwrapOrElseAsyncFromUndefinable as unwrapOrElseAsync } from './unwrapOrElseAsync.js';
+
+import { expectNotUndefined, unwrapUndefinable } from './Undefinable.js';
+
+/**
+ *  @deprecated
+ *  This is an alias for backward compatibility.
+ *  Please use {@link expectNotUndefined} instead.
+ */
+export const expect: typeof expectNotUndefined = expectNotUndefined;
+
+/**
+ *  @deprecated
+ *  This is an alias for backward compatibility.
+ *  Please use {@link unwrapUndefinable} instead.
+ */
+export const unwrap: typeof unwrapUndefinable = unwrapUndefinable;
